Seperate booklist to another file
