+++ src/components/book/Books.jsx
@@ -1,0 +1,11 @@
+import BookForm from './BookForm';
+import BookList from './BookList';
+
+const Books = () => (
+  <div>
+    <BookList />
+    <BookForm />
+  </div>
+);
+
+export default Books;
